requirements.txt
prusa.connect.sdk.printer>=0.8.1
py-gcode-metadata
pip>=22.2.0
bidict~=0.22.1
blinker~=1.5
extendparser~=0.3.1
jinja2-template-info~=0.2.4
lockfile~=0.12.2
packaging~=23.0
PoorWSGI~=2.5.0
pydantic==1.10.12 # to be kept pinned until https://github.com/pydantic/pydantic/issues/7689 is resolved
pyric~=0.1.6.3
python-daemon~=3.0.1
python-magic~=0.4.27
python-prctl~=1.8.1
PyTurboJPEG~=1.7.0
pyudev~=0.24.0
PyYAML~=6.0
requests>=2.31.0
sortedcontainers~=2.4.0
unidecode~=1.3.6
validators~=0.20.0
zeroconf~=0.47.4

# mqtt here
paho-mqtt>=2.0.0